Extract prepare step
